// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-word-type-selection-body.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingWordTypeScopeBody } from '../models/linking-word-type-scope-body';
export interface LinkingWordTypeSelectionBody {
  case: string | null;
  contextCode: string | null;
  kind: string | null;
  lemmaId: number | null;
  rootId: number | null;
  scope: LinkingWordTypeScopeBody | null;
  stemId: number | null;
  tashkeelWordId: number | null;
  tense: string | null;
  voice: string | null;
}